// WebApp/src/app/models/viewalbum.metadata.ts
export class ViewAlbumMetaData {
    AudioId: number;
    CoupleDetail: string;
    CreatedOn: Date;
    CustomerId: number;
    EAlbumId: number;
    Email: string;
    EventDate: Date;
    EventTitle: string;
    ModifiedOn: Date;
    ExpiredOn: Date;
    FullName:string;
    Mobile: string;
    Mp3Link:string;
    Mp3Title: string;
    PageType: string;
    PublishedOn: Date;
    Remarks:string;
    UserId: number;
    Status: string;
    UniqId: string;
    ParentId:string;
    IsDisplay: boolean;
}